[bessel-refactor] Use int64_t in CUDA bessel_gamma_sign and add xfail comments
Two minor convention fixes from a thorough audit pass:

1. bessel_gamma_sign in Math.cuh used 'long long' for the floor result
   while every other JIT string in the file (chebyshev_polynomial_*,
   etc.) uses int64_t. Switch to int64_t for consistency.

2. The test_compare_cpu expectedFailure decorators on
   special.modified_bessel_i and special.modified_bessel_k had no
   explanatory comment. Neighboring entries (chebyshev_polynomial_t,
   special.zeta) document why the comparison test fails. Add the same
   style of inline comment.

No behavior change.

Authored with assistance from Claude.

+++ aten/src/ATen/native/cuda/Math.cuh
@@ -2827,7 +2827,7 @@ const auto modified_bessel_i_string = modified_bessel_i0_string + modified_besse
         if (x >= T(0.0)) {
             return 1;
         }
-        long long n = static_cast<long long>(floor(-x + T(1.0)));
+        int64_t n = static_cast<int64_t>(floor(-x + T(1.0)));
         return (n % 2 == 0) ? 1 : -1;
     }
